files/requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
Flask==2.1.1
Flask_WTF==1.0.1
hed_utils==5.0.0
PyYAML==6.0
selenium==3.141.0
tabulate==0.8.9
WTForms==3.0.1
